Add skip_templates setting in config
The 'skip_templates' setting added in config can be used to specify
which files are to be skipped while generating the templates to
html files. The setting is in form of an array of unix-like wildcard
patterns (and NOT regex) in json.

As a default setting, 'base*.html' and '_*.html' has been added
to this list which means that when generate command is run, the
template files matching these patterns will not be generated.

In a typical scenario, base and partial templates should be named as
per these patterns.